hook moved outside and auth imported instead of direct way
