+++ frontend/src/pages/SellerDashboard.jsx
@@ -3,7 +3,7 @@
  * @author Thomas Noakes
  */
 
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 
 import {
@@ -24,61 +24,13 @@ import { useAuth } from "../context/AuthContext";
 import useSellerBundles from "../hooks/useSellerBundles";
 import { useSellerBundleReservations } from "../hooks/useSellerBundleReservations";
 import { useCollectReservation } from "../hooks/useCollectReservation";
+import { useSellerAllReservations } from "../hooks/useSellerAllReservations";
 
 import Card from "../components/Card";
 import Button from "../components/forms/Button";
 import FormInput from "../components/forms/FormInput";
 import SubmitButton from "../components/forms/SubmitButton";
 import Reservation from "../components/Reservation";
-
-const API_BASE_URL = import.meta.env.VITE_API_BASE_URL || "/api";
-
-function useAllSellerReservations(bundles) {
-    const [allReservations, setAllReservations] = useState([]);
-
-    useEffect(() => {
-        async function fetchAllReservations() {
-            if (!bundles || bundles.length === 0) {
-                return;
-            }
-
-            const token = localStorage.getItem("authToken");
-            if (!token) {
-                return;
-            }
-
-            const allRes = [];
-            try {
-                for (const bundle of bundles) {
-                    const response = await fetch(
-                        `${API_BASE_URL}/sellers/me/bundles/${bundle.bundle_id}/reservations`,
-                        {
-                            headers: {
-                                Authorization: `Bearer ${token}`,
-                            },
-                        },
-                    );
-                    if (response.ok) {
-                        const data = await response.json();
-                        allRes.push(
-                            ...data.map((r) => ({
-                                ...r,
-                                bundle_name: bundle.bundle_name,
-                            })),
-                        );
-                    }
-                }
-                setAllReservations(allRes);
-            } catch (err) {
-                console.error(err.message);
-            }
-        }
-
-        fetchAllReservations();
-    }, [bundles]);
-
-    return allReservations;
-}
 
 function CollectModal({ bundles, onClose }) {
     const [selectedBundleId, setSelectedBundleId] = useState("");
@@ -349,7 +301,7 @@ export default function SellerDashboard() {
     const navigate = useNavigate();
     const { bundles, loading } = useSellerBundles();
     const [showCollectModal, setShowCollectModal] = useState(false);
-    const allReservations = useAllSellerReservations(bundles);
+    const allReservations = useSellerAllReservations(bundles);
 
     if (userRole !== "seller") {
         return (
